requirements.txt
Flask==3.1.0
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.10
gunicorn==23.0.0
eventlet==0.37.0
python-dotenv==1.1.0
